# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: INOXYA BIJOUX API
  description: API REST pour la plateforme e-commerce INOXYA BIJOUX
  version: 1.0.0
  contact:
    name: Support INOXYA BIJOUX
    email: support@inoxya-bijoux.com
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://inoxya-bijoux.com/api
    description: Serveur de production
...
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Créer une commande
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Commande créée
        '400':
          description: Données invalides
        '429':
          description: Trop de requêtes
components:
  schemas:
...
    CreateOrderRequest:
      type: object
      required:
        - items
        - customer_name
        - customer_phone
        - customer_address
        - city
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              bijou_id:
                type: string
              pack_id:
                type: string
              quantity:
                type: number
        customer_name:
          type: string
        customer_phone:
          type: string
        customer_address:
          type: string
        city:
          type: string
        notes:
          type: string
...
security:
  - BearerAuth: []